+++ src/CloudService/custom/New-AzCloudServiceDiagnosticsExtension.ps1
@@ -19,11 +19,13 @@ Create a in-memory object for Diagnostics Extension
 Create a in-memory object for Diagnostics Extension
 
 .Outputs
-Microsoft.Azure.PowerShell.Cmdlets.CloudService.Models.Api20201001Preview.Extension
+Microsoft.Azure.PowerShell.Cmdlets.CloudService.Models.Api20210301.Extension
+.Link
+https://docs.microsoft.com/powershell/module/az.CloudService/New-AzCloudServiceDiagnosticsExtension
 #>
 
 function New-AzCloudServiceDiagnosticsExtension {
-  [OutputType('Microsoft.Azure.PowerShell.Cmdlets.CloudService.Models.Api20201001Preview.Extension')]
+  [OutputType('Microsoft.Azure.PowerShell.Cmdlets.CloudService.Models.Api20210301.Extension')]
   param(
     [Parameter(HelpMessage="Name of Diagnostics Extension.", Mandatory)]
     [string] $Name,