+++ src/Compute/Compute/VirtualMachine/Config/SetAzureVMSecurityProfile.cs
@@ -23,8 +23,8 @@ using Microsoft.Azure.Management.Compute.Models;
 
 namespace Microsoft.Azure.Commands.Compute
 {
-    [Cmdlet("Set", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "VMSecurityType"), OutputType(typeof(PSVirtualMachine))]
-    public class SetAzureVMSecurityType : Microsoft.Azure.Commands.ResourceManager.Common.AzureRMCmdlet
+    [Cmdlet("Set", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "VMSecurityProfile"), OutputType(typeof(PSVirtualMachine))]
+    public class SetAzureVMSecurityProfile : Microsoft.Azure.Commands.ResourceManager.Common.AzureRMCmdlet
     {
         [Alias("VMProfile")]
         [Parameter(