+++ src/Monitor/Monitor/ChangeLog.md
@@ -25,6 +25,7 @@
     - 'Get-AzDiagnosticSetting'
     - 'New-AzDiagnosticSetting'
     - 'Remove-AzDiagnosticSetting'
+* Supported `AutoMitigate` parameter in metric alert rule properties. The flag indicates whether the alert should be auto resolved or not.
 
 ## Version 2.4.0
 * Added cmdlets for data collection rules: