+++ src/Monitor/Monitor/ChangeLog.md
@@ -19,6 +19,13 @@
 -->
 
 ## Upcoming Release
+* Added cmdlet to get diagnostic setting categories for subscription
+    - `Get-AzSubscriptionDiagnosticSettingCategory`
+* Supported subscription diagnostic setting operations with new parameter: SubscriptionId
+    - 'Get-AzDiagnosticSetting'
+    - 'New-AzDiagnosticSetting'
+    - 'Remove-AzDiagnosticSetting'
+* Supported `AutoMitigate` parameter in metric alert rule properties. The flag indicates whether the alert should be auto resolved or not.
 
 ## Version 2.4.0
 * Added cmdlets for data collection rules: